Make two small optimizations
For comp_aware_bin_split_2, if the discovered component ends at the end
of the searched text, then there is no possibility to discover a new
component, so skip the check to see if we have found all components yet.
(We already did this optimization in comp_aware_bin_split.)

For both comp_aware_bin_split and comp_aware_bin_split_2, if we are
(e.g.) searching a text ABCDEFG and we find one component BCDE and find
another component that begins at C, then we know that the second
component must extend at least as long to CDEF, so don't check anything
shorter.  In general, if we find a component that begins inside of
another, then it must extend to at least the end of the other plus one.
Else it would be a substring of the other, which would be a more general
matching rule than the one that we already found.